requirements.txt
streamlit==1.41.1
google-generativeai==0.8.3